Implement `IAsyncMigrationProcessor` in `ProcessorBase`
To avoid a lot of code duplication, all classes, that implement an
synchronous database access interface should implement the asynchronous
counterpart as well. `ProcessorBase` is the base class for most (all?)
implementations of the `IMigrationProcessor` interface.

If the synchronous equivalent of a method was implemented `abstract` the
asynchronous method is implemented `virtual` and will inline a call to
the synchronous equivalent. Initially i thought about throwing an
`NotSupportedExcetion`, because there are reasons agains inlining the
calls, as i did and using `Task.Run()` to achieve pseudo asynchroniness.
There is a chanche in both cases, that at ThreadPoolThread is blocked,
without the called being aware of that. But when further implementing
the asynchrounous methods it became clear, that some connections have
support for some asynchronous operations. Most notable are transactions,
which cannot generally be started asynchronously.

Because i didn't want to let async go for a connection, because maybe
one aspect was not supported, i changed my mind and implemented the
asynchronous methods as they are now.

When an asynchronous call is redirected to an other asynchronous method,
simply the task of the other method is returned, to reduce the amount of
created tasks.