Add async migration support for SqlServer
Implement the async overloads of `ProcessorBase` in
`SqlServerProcessor`. To reduce the amount of duplicate code, i
sometimes used the _sync_-PAttern used by Stephen Cleary in AsyncEx for
methods that can be run synchronous as well as asynchronous. The idea
is, that when the `sync` parameter is true no execution path will reach
an await.